+++ gpjax/__init__.py
@@ -13,12 +13,14 @@
 # limitations under the License.
 # ==============================================================================
 
+from . import _version
 from .abstractions import fit, fit_batches, fit_natgrads
 from .gps import Prior, construct_posterior
 from .kernels import *
 from .likelihoods import Bernoulli, Gaussian
 from .mean_functions import Constant, Zero
-from .parameters import constrain, copy_dict_structure, initialise, unconstrain
+from .params import constrain, copy_dict_structure, initialise, unconstrain
+from .types import Dataset
 from .variational_families import (
     CollapsedVariationalGaussian,
     ExpectationVariationalGaussian,
@@ -26,9 +28,7 @@ from .variational_families import (
     VariationalGaussian,
     WhitenedVariationalGaussian,
 )
-from .types import Dataset
 from .variational_inference import CollapsedVI, StochasticVI
-from . import _version
 
 __version__ = _version.get_versions()["version"]
 __license__ = "MIT"
