+++ gpjax/kernels/nonstationary/linear.py
@@ -20,7 +20,7 @@ from ..base import AbstractKernel
 
 from dataclasses import dataclass
 from jaxtyping import Array, Float
-from mytree import param_field, Softplus
+from ...parameters import param_field, Softplus
 
 @dataclass
 class Linear(AbstractKernel):
