+++ gpjax/mean_functions.py
@@ -16,12 +16,11 @@
 import abc
 from typing import Dict, Optional
 
+import deprecation
 import jax.numpy as jnp
 from jax.random import KeyArray
 from jaxtyping import Array, Float
 from jaxutils import PyTree
-
-import deprecation
 
 
 class AbstractMeanFunction(PyTree):
