+++ tests/test_gaussian_distribution.py
@@ -23,10 +23,9 @@ from jax.config import config
 # Enable Float64 for more stable matrix inversions.
 config.update("jax_enable_x64", True)
 
+from gpjax.gaussian_distribution import GaussianDistribution
 from gpjax.linops.dense_linear_operator import DenseLinearOperator
 from gpjax.linops.diagonal_linear_operator import DiagonalLinearOperator
-
-from gpjax.gaussian_distribution import GaussianDistribution
 
 _key = jr.PRNGKey(seed=42)
