+++ tests/test_likelihoods.py
@@ -20,8 +20,8 @@ import jax.numpy as jnp
 import jax.random as jr
 import numpy as np
 import pytest
-from jax.random import KeyArray
 from jax.config import config
+from jax.random import KeyArray
 from jaxtyping import Array, Float
 
 from gpjax.likelihoods import (
@@ -32,7 +32,6 @@ from gpjax.likelihoods import (
     NonConjugate,
     inv_probit,
 )
-
 
 # Enable Float64 for more stable matrix inversions.
 config.update("jax_enable_x64", True)
